Refactor MonacoEditor to use next/dynamic for client-side loading
- Extracted editor implementation to `components/MonacoEditorClient.tsx`
- Updated `components/MonacoEditor.tsx` to dynamically import the client component with `{ ssr: false }`
- Added `MonacoEditorProps` to `components/types.ts` for shared typing
- Removed manual `isClient` state and `useEffect` check, resolving the `setState` in `useEffect` warning
- Preserved loading state and container styling
