+++ components/types.ts
@@ -4,3 +4,9 @@ export interface BlobEntry {
   url: string;
   snippet: string;
 }
+
+export interface MonacoEditorProps {
+  value: string;
+  onChange: (value: string) => void;
+  className?: string;
+}
